fix(youtube/queue): treat a missing job row as terminal, canonicalise channel handles in lib
